Skip optimization skipping relation tracking of test files
Should improve the situation in #19 - I wasn't able to observe any performance
regressions from doing the change, and I guess it's simpler if we don't need to
handle that extra case.

If we decide to go with this solution we should remove the entire if branch
instead of just commenting it out.