Remember the active prompt config & require two Esc presses to exit.
Set a flag at the point where a prompt is applied via its Enter key.
Thus, when registering an Esc press while within a prompt first it
is displayed. With this behaviour we can now also restore the last
argument that was used. A second Esc press clears the mode and args.