Add discovered files to a queue that is then flushed at test run time.
As files are discovered the test timer is pushed back. In addition
to doing this, this commit also postpones processing added files
until we come to run the tests.

What this enables is only recalculating the files that would be run
by mocha once, apply this update wholesale to the SourceGraph and
then allow addFileFromPath() to take care of the is testFile check.